+++ fabric-item-api-v1-bridge/src/main/java/net/fabricmc/fabric/impl/item/ItemExtensions.java
@@ -2,6 +2,7 @@ package net.fabricmc.fabric.impl.item;
 
 import net.fabricmc.fabric.api.item.v1.CustomDamageHandler;
 import net.fabricmc.fabric.api.item.v1.EquipmentSlotProvider;
+import net.minecraft.core.component.DataComponentMap;
 import org.jetbrains.annotations.Nullable;
 
 public interface ItemExtensions {
@@ -9,4 +10,5 @@ public interface ItemExtensions {
     void fabric_setEquipmentSlotProvider(@Nullable EquipmentSlotProvider provider);
     @Nullable CustomDamageHandler fabric_getCustomDamageHandler();
     void fabric_setCustomDamageHandler(@Nullable CustomDamageHandler handler);
+    void fabric_setDefaultComponents(DataComponentMap components);
 }